pyproject.toml
[project]
name = "diskdantic"
version = "0.1.0"
description = "Disk-backed collections powered by Pydantic models."
readme = "README.md"
authors = [
    { name = "koaning", email = "vincentwarmerdam@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.12.4",
    "pyyaml>=6.0.3",
    "orjson>=3.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "ruff>=0.8.4",
    "pre-commit>=4.0.1",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "UP046",  # Generic[T] syntax - keeping for compatibility
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
